Feature/pedidos sincronizados (#10)
* se guarda la direccion del usuario al hacer la compra, aparte de eso. Al finalizar la compra la vista de pedidos se actualiza para mostrar el nuevo pedido y el stock se actualiza.

* feat(pedidos): DTOs y servicios consistentes, subtotal calculado por BD, controllers limpios

+++ Back-end/pedidos-api-descansosdelrecuerdospa/src/main/java/com/descansos_del_recuerdo_spa/pedidos/repositories/DetallePedidoRepository.java
@@ -6,11 +6,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Integer> {
+public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Long> {
 
     // 🔹 Buscar detalles por pedido
-    List<DetallePedido> findByPedido_Id(Integer pedidoId);
+    List<DetallePedido> findByPedido_Id(Long pedidoId);
 
     // 🔹 Buscar detalles por urna
-    List<DetallePedido> findByUrnaId(Integer urnaId);
+    List<DetallePedido> findByUrnaId(Long urnaId);
 }
