+++ Back-end/pedidos-api-descansosdelrecuerdospa/src/main/java/com/descansos_del_recuerdo_spa/pedidos/repositories/EstadoPedidoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Integer> {
+public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Long> {
 
     // 🔹 Buscar estado por nombre (útil para lógica interna o inicialización)
     EstadoPedido findByNombre(String nombre);
