+++ Back-end/pedidos-api-descansosdelrecuerdospa/src/main/java/com/descansos_del_recuerdo_spa/pedidos/services/EstadoPedidoService.java
@@ -6,6 +6,6 @@ import java.util.List;
 public interface EstadoPedidoService {
 
     List<EstadoPedido> listarEstados();
-    EstadoPedido obtenerPorId(Integer id);
+    EstadoPedido obtenerPorId(Long id);
     EstadoPedido obtenerPorNombre(String nombre);
 }
